+++ app/cart/biz/model/cartData.go
@@ -2,6 +2,6 @@ package model
 
 type Cart struct {
 	UserID    uint32
-	ProductID uint32
+	ProductID int64
 	Quantity  int32
 }
